+++ tests/testthat/test_bnpsd.R
@@ -525,6 +525,13 @@ test_that("draw_p_anc is in range", {
     expect_equal(length(p_anc), m_loci)
     expect_true(all(p_anc >= 0)) # all are non-negative
     expect_true(all(p_anc <= 1)) # all are smaller or equal than 1
+
+    # repeat for Beta version
+    beta <- 0.01
+    p_anc <- draw_p_anc(m_loci, beta = beta)
+    expect_equal(length(p_anc), m_loci)
+    expect_true(all(p_anc >= 0)) # all are non-negative
+    expect_true(all(p_anc <= 1)) # all are smaller or equal than 1
 })
 
 test_that("draw_p_subpops is in range", {
@@ -692,6 +699,36 @@ test_that("draw_all_admix works well", {
     expect_equal(ncol(p_subpops), k_subpops)
     expect_true(all(p_subpops >= 0)) # all are non-negative
     expect_true(all(p_subpops <= 1)) # all are smaller or equal than 1
+    
+    # test p_anc
+    expect_equal(length(p_anc), m_loci)
+    expect_true(all(p_anc >= 0)) # all are non-negative
+    expect_true(all(p_anc <= 1)) # all are smaller or equal than 1
+})
+
+test_that("draw_all_admix beta works well", {
+    m_loci <- 10
+    beta <- 0.01
+    inbr_subpops <- c(0.1, 0.2, 0.3)
+    k_subpops <- length(inbr_subpops)
+    admix_proportions <- diag(rep.int(1, k_subpops)) # island model for test...
+    # repeat so we have multiple people per island
+    admix_proportions <- rbind(admix_proportions, admix_proportions, admix_proportions)
+    n_ind <- nrow(admix_proportions) # number of individuals (3*k_subpops)
+
+    # run draw_all_admix
+    # only test default (p_ind and p_subpops not returned)
+    out <- draw_all_admix(admix_proportions, inbr_subpops, m_loci, beta = beta)
+    expect_equal( names(out), draw_all_admix_names_ret_default )
+
+    X <- out$X # genotypes
+    p_anc <- out$p_anc # Ancestral AFs
+    
+    # test X
+    expect_equal(nrow(X), m_loci)
+    expect_equal(ncol(X), n_ind)
+    expect_true(all(X %in% c(0, 1, 2))) # only three values allowed!
+    expect_true(!any(fixed_loci(X))) # we don't expect any loci to be fixed
     
     # test p_anc
     expect_equal(length(p_anc), m_loci)